% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiTest.R
\name{chi_square_test}
\alias{chi_square_test}
\title{Uji Chi-Square untuk Data Diskrit}
\usage{
chi_square_test(data, dist = "binomial", params = NULL)
}
\arguments{
\item{data}{Vector data diskrit.}

\item{dist}{Nama distribusi teoretik: "bernoulli", "binomial", "poisson", atau "geometric".
Jika NULL, maka digunakan "binomial" secara default.}

\item{params}{List parameter distribusi. Jika NULL, akan diestimasi dari data.}
}
\value{
Objek dari \code{chisq.test}
}
\description{
Fungsi ini melakukan uji Chi-Square goodness-of-fit untuk distribusi diskrit
(bernoulli, binomial, poisson, geometric) dengan parameter bisa ditentukan
atau diestimasi otomatis dari data.
}
\examples{
x <- simulate_data(100, "binomial", list(2, 0.5))
chi_square_test(x)  # otomatis deteksi distribusi dan estimasi parameter
chi_square_test(x, "binomial", list(size = 2, prob = 0.5))

x <- simulate_data(100, "poisson", list(3))
chi_square_test(x, "poisson")

}
