% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiTest.R
\name{chisq_test_manual}
\alias{chisq_test_manual}
\title{Chi-Square GOF Test Manual (Setara chisq.test bawaan R)}
\usage{
chisq_test_manual(data, dist, params)
}
\arguments{
\item{data}{Vector numerik diskrit (integer).}

\item{dist}{Distribusi: "binomial", "poisson", "geometric", "nbinom"}

\item{params}{List parameter distribusi. Lihat dokumentasi masing-masing.}
}
\value{
List hasil uji: statistik, df, p-value, observed, expected
}
\description{
Chi-Square GOF Test Manual (Setara chisq.test bawaan R)
}
