% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelompok10-package.R
\docType{package}
\name{kelompok10-package}
\alias{kelompok10}
\alias{kelompok10-package}
\title{kelompok10: Goodness-of-Fit Tests for Simulated Data}
\description{
Package to perform chi-square and Kolmogorov-Smirnov tests on simulated data.
}
\keyword{internal}
