% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksTest.R
\name{ks_test}
\alias{ks_test}
\title{Uji Kolmogorov-Smirnov Manual}
\usage{
ks_test(data, dist = "norm", params = NULL)
}
\arguments{
\item{data}{Data numerik vektor.}

\item{dist}{Karakter nama distribusi teoretik: "norm", "exp", "gamma", atau "beta".}

\item{params}{List parameter distribusi. Misalnya, list(mean = 0, sd = 1) untuk normal.}
}
\value{
List dengan nilai D, Dplus, Dminus, dan p-value (aproksimasi).
}
\description{
Melakukan uji KS satu sampel secara manual terhadap distribusi teoretik (normal, eksponensial, dll).
}
\examples{
x <- rnorm(30)
ks_test(x, "norm", list(mean = 0, sd = 1))
}
