% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksTest.R
\name{ks_test}
\alias{ks_test}
\title{Uji Kolmogorov-Smirnov untuk Data Kontinu}
\usage{
ks_test(data, dist = "normal", params = NULL)
}
\arguments{
\item{data}{Vector numerik. Data hasil pengamatan yang akan diuji.}

\item{dist}{Karakter. Nama distribusi teoretik: "normal", "exponential", "gamma", atau "beta".}

\item{params}{List parameter distribusi. Jika NULL, akan dihitung otomatis dari data.}
}
\value{
Objek hasil dari fungsi \code{ks.test}.
}
\description{
Fungsi ini melakukan uji Kolmogorov-Smirnov untuk menguji apakah data berasal
dari distribusi kontinu teoretik tertentu (normal, eksponensial, gamma, atau beta).
}
\examples{
x <- simulate_data(100, "beta", list(2, 5))
ks_test(x, "beta")  # estimasi otomatis shape1 dan shape2

x <- simulate_data(100, "normal", list(0, 1))
ks_test(x)  # default: normal

}
