% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksTest.R
\name{ks_test}
\alias{ks_test}
\title{Uji Kolmogorov-Smirnov untuk Data Kontinu}
\usage{
ks_test(data, dist = "normal", params = NULL, auto_scale = FALSE)
}
\arguments{
\item{data}{Vector numerik. Data hasil pengamatan yang akan diuji.}

\item{dist}{Karakter. Nama distribusi teoretik, salah satu dari:
`"normal"`, `"exponential"`, `"gamma"`, `"beta"`.}

\item{params}{List berisi parameter distribusi teoretik. Jika `NULL`, akan diestimasi dari data.}

\item{auto_scale}{Logical. Jika `TRUE` dan distribusi adalah "beta", data di luar (0,1) akan diskalakan otomatis.}
}
\value{
Objek hasil dari fungsi \code{ks.test}.
}
\description{
Fungsi ini melakukan uji Kolmogorov-Smirnov untuk menguji apakah data berasal
dari distribusi kontinu teoretik tertentu (normal, eksponensial, gamma, atau beta).
}
\examples{
x <- rnorm(100)
ks_test(x, "normal")

y <- runif(100)
ks_test(y, "beta")  # Tidak perlu penskalaan

z <- rnorm(100, mean = 5, sd = 2)
ks_test(z, "beta", auto_scale = TRUE)  # Otomatis diskalakan ke (0,1)

}
